% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Econo.R
\name{bgtest}
\alias{bgtest}
\title{Breusch-Godfrey Test}
\usage{
bgtest(
  formula,
  order = 1,
  order.by = NULL,
  type = c("Chisq", "F"),
  data = list(),
  fill = 0
)
}
\arguments{
\item{formula}{a symbolic description for the model to be tested (or a fitted "lm" object).}

\item{order}{integer. maximal order of serial correlation to be tested.}

\item{order.by}{Either a vector z or a formula with a single explanatory variable like ~ z. The observations in the model are ordered by the size of z. If set to NULL (the default) the observations are assumed to be ordered (e.g., a time series).}

\item{type}{the type of test statistic to be returned. Either "Chisq" for the Chi-squared test statistic or "F" for the F test statistic}

\item{data}{an optional data frame containing the variables in the model. By default the variables are taken from the environment which bgtest is called from.}

\item{fill}{starting values for the lagged residuals in the auxiliary regression. By default 0 but can also be set to NA.}
}
\description{
bgtest performs the Breusch-Godfrey test for higher-order serial correlation
}
\details{
Under H_0 the test statistic is asymptotically Chi-squared with degrees of freedom as given in parameter. If type is set to "F" the function returns a finite sample version of the test statistic, employing an F distribution with degrees of freedom as given in parameter.
}
\references{
Inspired by Gold Medalist Dr Shahla Faisal who teach me in GC University Faisalabad
}
\author{
Syed Hammad Raza
}
