% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Econo.R
\name{bptest}
\alias{bptest}
\title{Breusch-Pagan Tes}
\usage{
bptest(formula, varformula = NULL, data = list(), weights = NULL)
}
\arguments{
\item{formula}{symbolic description for the model to be tested (or a fitted "lm" object)}

\item{varformula}{a formula describing only the potential explanatory variables for the variance (no dependent variable needed). By default the same explanatory variables are taken as in the main regression model.}

\item{data}{an optional data frame containing the variables in the model. By default the variables are taken from the environment which bptest is called from.}

\item{weights}{an optional vector of weights to be used in the model.}
}
\description{
Performs the Breusch-Pagan test against heteroskedasticity.
}
\details{
The Breusch-Pagan test fits a linear regression model to the residuals of a linear regression model (by default the same explanatory variables are taken as in the main regression model) and rejects if too much of the variance is explained by the additional explanatory variables
}
\references{
Inspired by Gold Medalist Dr Shahla Faisal who teach me in GC University Faisalabad
}
\author{
Syed Hammad Raza
}
