% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Econo.R
\name{CI}
\alias{CI}
\title{Condition Index(CI) in R}
\usage{
CI(m)
}
\arguments{
\item{m}{is a matrix or data.frame}
}
\description{
If the Condition Index is between 10 and 30, there is moderate to strong multicollinearity and if it exceeds 30 there is severe multicollinearity
}
\references{
Inspired by Gold Medalist Dr Shahla Faisal who teach me in GC University Faisalabad
}
\author{
Syed Hammad Raza
}
