% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Econo.R
\name{dwtest}
\alias{dwtest}
\title{Durbin-Watson Test}
\usage{
dwtest(
  formula,
  order.by = NULL,
  alternative = c("greater", "two.sided", "less"),
  iterations = 15,
  exact = NULL,
  tol = 1e-10,
  data = list()
)
}
\arguments{
\item{formula}{a symbolic description for the model to be tested (or a fitted "lm" object).}

\item{order.by}{Either a vector z or a formula with a single explanatory variable like ~ z. The observations in the model are ordered by the size of z. If set to NULL (the default) the observations are assumed to be ordered (e.g., a time series).}

\item{alternative}{a character string specifying the alternative hypothesis.}

\item{iterations}{an integer specifying the number of iterations when calculating the p-value with the "pan" algorithm.}

\item{exact}{logical. If set to FALSE a normal approximation will be used to compute the p value, if TRUE the "pan" algorithm is used. The default is to use "pan" if the sample size is < 100.}

\item{tol}{tolerance. Eigenvalues computed have to be greater than tol to be treated as non-zero.}

\item{data}{an optional data frame containing the variables in the model. By default the variables are taken from the environment which dwtest is called from.}
}
\description{
Performs the Durbin-Watson test for autocorrelation of disturbances.
}
\details{
The Durbin-Watson test has the null hypothesis that the autocorrelation of the disturbances is 0. It is possible to test against the alternative that it is greater than, not equal to, or less than 0, respectively. This can be specified by the alternative argument.
}
\references{
Inspired by Gold Medalist Dr Shahla Faisal who teach me in GC University Faisalabad
}
\author{
Syed Hammad Raza
}
