% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Econo.R
\name{gqtest}
\alias{gqtest}
\title{Goldfeld-Quandt Tes}
\usage{
gqtest(
  formula,
  point = 0.5,
  fraction = 0,
  alternative = c("greater", "two.sided", "less"),
  order.by = NULL,
  data = list()
)
}
\arguments{
\item{formula}{a symbolic description for the model to be tested (or a fitted "lm" object).}

\item{point}{numerical. If point is smaller than 1 it is interpreted as percentages of data, i.e. n*point is taken to be the (potential) breakpoint in the variances, if n is the number of observations in the model. If point is greater than 1 it is interpreted to be the index of the breakpoint.}

\item{fraction}{numerical. The number of central observations to be omitted. If fraction is smaller than 1, it is chosen to be fraction*n if n is the number of observations in the model.}

\item{alternative}{a character string specifying the alternative hypothesis. The default is to test for increasing variances.}

\item{order.by}{Either a vector z or a formula with a single explanatory variable like ~ z. The observations in the model are ordered by the size of z. If set to NULL (the default) the observations are assumed to be ordered (e.g., a time series).}

\item{data}{an optional data frame containing the variables in the model. By default the variables are taken from the environment which gqtest is called from.}
}
\description{
Goldfeld-Quandt test against heteroscedasticity
}
\details{
The Goldfeld-Quandt test compares the variances of two submodels divided by a specified breakpoint and rejects if the variances differ.
}
\references{
Inspired by Gold Medalist Dr Shahla Faisal who teach me in GC University Faisalabad
}
\author{
Syed Hammad Raza
}
