% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Econo.R
\name{vif}
\alias{vif}
\title{Variance Inflation Method in R}
\usage{
vif(mod, ...)
}
\arguments{
\item{mod}{is a Regression Model which we can fit by lm() command}
}
\description{
When x  is nearly linearly independent on subset of remaining regressor. If R^2 is nearly 1 it mean VIF is large. As a rule of thumb high value if VIF of variable is exceed 10, mean multicollinearity is exit.
}
\references{
Inspired by Gold Medalist Dr Shahla Faisal who teach me in GC University Faisalabad
}
\author{
Syed Hammad Raza
}
