% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Econo.R
\name{white.test}
\alias{white.test}
\title{White's Test for Heteroskedasticity in a Linear Regression Mode}
\usage{
white.test(lmobj, squares.only = FALSE)
}
\arguments{
\item{lmobj}{Either an object of class "lm" (e.g., generated by lm), or a list of two objects: a response vector and a design matrix. The objects are assumed to be in that order, unless they are given the names "X" and "y" to distinguish them. The design matrix passed in a list must begin with a column of ones if an intercept is to be included in the linear model. The design matrix passed in a list should not contain factors, as all columns are treated 'as is'. For tests that use ordinary least squares residuals, one can also pass a vector of residuals in the list, which should either be the third object or be named "e".}

\item{squares.only}{A logical. Should two-way interactions between explanatory variables be included in the auxiliary regression? Defaults to FALSE, since when interaction terms are present the test is not a pure test of heteroskedasticity but also of model specification.}
}
\description{
This function implements the popular method of White (1980) for testing for heteroskedasticity in a linear regression model.
}
\details{
White's Test entails fitting an auxiliary regression model in which the response variable is the vector of squared residuals from the original model and the design matrix includes the original explanatory variables, their squares, and (optionally) their two-way interactions. The test statistic is the number of observations multiplied by the coefficient of determination from the auxiliary regression model
}
\references{
Inspired by Gold Medalist Dr Shahla Faisal who teach me in GC University Faisalabad
}
\author{
Syed Hammad Raza
}
